Initial commit: add main.py and resources (dir)
This script was previously hosted on

https://github.com/ningyuansg/Twitter-Weather-Bot

In its previous form (Twitter-Weather-Bot), sensitive information
i.e. twitter access tokens were included within the script as
string values. Each push to github thus required the removal of
those values each time. As a result, commits to that repository
were rare.

In the current iteration (nimbus), the above information is stored
in an external json file (config.json). This file is not tracked.
As a result, commits are now easier and will thus be more frequent.